Moved useRenditionService2(...) to the old RenditionService as we will need to call it from multiple onContentUpdates but only as long as there is the old service.
