+++ src/main/resources/alfresco/rendition-services-context.xml
@@ -49,6 +49,7 @@
       <property name="renditionDefinitionPersister" ref="renditionDefinitionPersister" />
       <property name="renditionPreventionRegistry" ref="renditionPreventionRegistry" />
       <property name="policyComponent" ref="policyComponent" />
+       <property name="renditionService2" ref="RenditionService2"/>
       <property name="knownCancellableActionTypes">
         <list>
             <bean id="org.alfresco.repo.rendition.executer.ImageRenderingEngine.NAME"
@@ -211,7 +212,6 @@
        <property name="nodeService" ref="NodeService"/>
        <property name="policyComponent" ref="policyComponent"/>
        <property name="renditionService" ref="RenditionService"/>
-       <property name="renditionService2" ref="RenditionService2"/>
    </bean>
 
    <!-- This action is used as a compensating action in the event of a failure to update a rendition -->
